requirements.txt
strands-agents==1.0.0
strands-agents-tools==0.2.1
yt-dlp
boto3